requirements.txt
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
